# swagger-api-documentation.yaml
openapi: 3.0.3
info:
  title: Agentic Itinerary Planner API
  description: |
    Comprehensive API for AI-powered travel itinerary planning and management.
    
    This API provides endpoints for:
    - Creating and managing travel itineraries
    - Real-time agent progress tracking via Server-Sent Events
    - Travel planning tools (weather, packing lists, photo spots, etc.)
    - Booking and payment processing
    - Health monitoring and testing
    
    All endpoints support CORS and work without authentication in development mode.
  version: 1.0.0
  contact:
    name: Agentic Itinerary Planner Team
    email: support@agenticitinerary.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.agenticitinerary.com/api/v1
    description: Production server
...
paths:
...
  /itineraries:
    post:
      tags:
        - Itineraries
      summary: Create a new itinerary
      description: Creates a new travel itinerary with AI agent processing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItineraryReq'
            example:
              destination: "Barcelona, Spain"
              startDate: "2025-06-01"
              endDate: "2025-06-04"
              party:
                adults: 2
                children: 1
                infants: 0
                rooms: 1
              budgetTier: "mid-range"
              interests: ["culture", "architecture", "food"]
              constraints: ["familyFriendly", "wheelchairAccessible"]
              language: "en"
      responses:
        '200':
          description: Itinerary created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Itineraries
      summary: Get all itineraries
      description: Retrieves a list of all itineraries
      responses:
        '200':
          description: List of itineraries retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItineraryDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /itineraries/{id}:
    get:
      tags:
        - Itineraries
      summary: Get itinerary by ID
      description: Retrieves a specific itinerary by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Itinerary ID
          schema:
            type: string
            example: "3"
      responses:
        '200':
          description: Itinerary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDto'
        '404':
          description: Itinerary not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Itineraries
      summary: Delete itinerary
      description: Deletes a specific itinerary by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Itinerary ID
          schema:
            type: string
            example: "3"
      responses:
        '204':
          description: Itinerary deleted successfully
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Agent Endpoints
  /agents/stream:
    get:
      tags:
        - Agents
      summary: Stream agent events (SSE)
      description: |
        Establishes a Server-Sent Events connection to receive real-time updates
        about AI agent progress for itinerary generation.
      parameters:
        - name: itineraryId
          in: query
          required: true
          description: Itinerary ID to track
          schema:
            type: string
            example: "3"
      responses:
        '200':
          description: SSE connection established
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: connected
                  data: Connected to agent stream for itinerary: 3
                  
                  event: agent-progress
                  data: {"kind":"planner","status":"running","progress":45,"message":"Generating activities..."}
...
components:
  schemas:
...
    CreateItineraryReq:
      type: object
      required:
        - destination
        - startDate
        - endDate
        - party
        - budgetTier
      properties:
        destination:
          type: string
          description: Travel destination
          example: "Barcelona, Spain"
        startDate:
          type: string
          format: date
          description: Trip start date
          example: "2025-06-01"
        endDate:
          type: string
          format: date
          description: Trip end date
          example: "2025-06-04"
        party:
          $ref: '#/components/schemas/PartyDto'
        budgetTier:
          type: string
          description: Budget tier for the trip
          enum: [economy, mid-range, luxury]
          example: "mid-range"
        interests:
          type: array
          items:
            type: string
          description: List of travel interests
          example: ["culture", "architecture", "food"]
        constraints:
          type: array
          items:
            type: string
          description: List of travel constraints
          example: ["familyFriendly", "wheelchairAccessible"]
        language:
          type: string
          description: Preferred language
          default: "en"
          example: "en"
    PartyDto:
      type: object
      required:
        - adults
      properties:
        adults:
          type: integer
          minimum: 1
          description: Number of adults
          example: 2
        children:
          type: integer
          minimum: 0
          default: 0
          description: Number of children
          example: 1
        infants:
          type: integer
          minimum: 0
          default: 0
          description: Number of infants
          example: 0
        rooms:
          type: integer
          minimum: 1
          default: 1
          description: Number of rooms needed
          example: 1

    # Response DTOs
    ItineraryDto:
      type: object
      properties:
        id:
          type: string
          description: Unique itinerary identifier
          example: "3"
        destination:
          type: string
          description: Travel destination
          example: "Barcelona, Spain"
        startDate:
          type: string
          format: date
          description: Trip start date
          example: "2025-06-01"
        endDate:
          type: string
          format: date
          description: Trip end date
          example: "2025-06-04"
        party:
          $ref: '#/components/schemas/PartyDto'
        budgetTier:
          type: string
          description: Budget tier for the trip
          example: "mid-range"
        interests:
          type: array
          items:
            type: string
          description: List of travel interests
          example: ["culture", "architecture", "food"]
        constraints:
          type: array
          items:
            type: string
          description: List of travel constraints
          example: ["familyFriendly", "wheelchairAccessible"]
        language:
          type: string
          description: Preferred language
          example: "en"
        summary:
          type: string
          description: Trip summary
          example: "Your personalized itinerary for Barcelona, Spain"
        map:
          type: object
          description: Map data (not implemented yet)
          nullable: true
        days:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDayDto'
          description: Daily itinerary details
        agentResults:
          type: object
          description: AI agent processing results
          nullable: true
        status:
          type: string
          description: Itinerary status
          enum: [draft, generating, completed, failed]
          example: "generating"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-09-19T15:21:44.259271500Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-09-19T15:21:44.259271500Z"
        isPublic:
          type: boolean
          description: Whether itinerary is public
          example: false
        shareToken:
          type: string
          description: Share token for public access
          nullable: true
    ItineraryDayDto:
      type: object
      properties:
        day:
          type: integer
          description: Day number in the itinerary
          example: 1
        date:
          type: string
          format: date
          description: Date for this day
          example: "2025-06-01"
        location:
          type: string
          description: Primary location for the day
          example: "Barcelona, Spain"
        activities:
          type: array
          items:
            $ref: '#/components/schemas/ActivityDto'
          description: Activities planned for this day
        accommodation:
          $ref: '#/components/schemas/AccommodationDto'
          description: Accommodation for this day
        transportation:
          type: array
          items:
            $ref: '#/components/schemas/TransportationDto'
          description: Transportation for this day
        meals:
          type: array
          items:
            $ref: '#/components/schemas/MealDto'
          description: Meals planned for this day
        notes:
          type: string
          description: Additional notes for this day
    ActivityDto:
      type: object
      properties:
        name:
          type: string
          description: Activity name
          example: "Sagrada Familia Tour"
        description:
          type: string
          description: Activity description
          example: "Guided tour of Gaudi's masterpiece"
        location:
          $ref: '#/components/schemas/LocationDto'
        startTime:
          type: string
          description: Activity start time
          example: "09:00"
        endTime:
          type: string
          description: Activity end time
          example: "11:00"
        duration:
          type: string
          description: Activity duration
          example: "120"
        category:
          type: string
          description: Activity category
          example: "sightseeing"
        price:
          $ref: '#/components/schemas/PriceDto'
        bookingRequired:
          type: boolean
          description: Whether booking is required
          example: true
        bookingUrl:
          type: string
          description: URL for booking
          example: "https://example.com/book"
        tips:
          type: string
          description: Tips for this activity
    AccommodationDto:
      type: object
      properties:
        name:
          type: string
          description: Accommodation name
          example: "Hotel Casa Fuster"
        type:
          type: string
          description: Accommodation type
          example: "hotel"
        location:
          $ref: '#/components/schemas/LocationDto'
        checkIn:
          type: string
          format: date
          description: Check-in date
          example: "2025-06-01"
        checkOut:
          type: string
          format: date
          description: Check-out date
          example: "2025-06-04"
        price:
          $ref: '#/components/schemas/PriceDto'
        rating:
          type: number
          format: double
          description: Accommodation rating
          example: 4.5
        amenities:
          type: array
          items:
            type: string
          description: List of amenities
          example: ["wifi", "pool", "gym"]
        bookingUrl:
          type: string
          description: URL for booking
          example: "https://example.com/book"
    TransportationDto:
      type: object
      properties:
        mode:
          type: string
          description: Transportation mode
          example: "flight"
        from:
          $ref: '#/components/schemas/LocationDto'
        to:
          $ref: '#/components/schemas/LocationDto'
        departureTime:
          type: string
          description: Departure time
          example: "14:30"
        arrivalTime:
          type: string
          description: Arrival time
          example: "16:45"
        duration:
          type: string
          description: Journey duration
          example: "135"
        price:
          $ref: '#/components/schemas/PriceDto'
        provider:
          type: string
          description: Transportation provider
          example: "Iberia"
        bookingUrl:
          type: string
          description: URL for booking
          example: "https://example.com/book"
        notes:
          type: string
          description: Additional notes
    MealDto:
      type: object
      properties:
        type:
          type: string
          description: Meal type
          example: "dinner"
        name:
          type: string
          description: Meal name
          example: "Paella Valenciana"
        restaurant:
          type: string
          description: Restaurant name
          example: "Restaurante 7 Portes"
        location:
          $ref: '#/components/schemas/LocationDto'
        time:
          type: string
          description: Meal time
          example: "20:00"
        price:
          $ref: '#/components/schemas/PriceDto'
        cuisine:
          type: string
          description: Cuisine type
          example: "Spanish"
        notes:
          type: string
          description: Additional notes
    LocationDto:
      type: object
      properties:
        name:
          type: string
          description: Location name
          example: "Sagrada Familia"
        address:
          type: string
          description: Full address
          example: "Carrer de Mallorca, 401, 08013 Barcelona, Spain"
        lat:
          type: number
          format: double
          description: Latitude coordinate
          example: 41.4036
        lng:
          type: number
          format: double
          description: Longitude coordinate
          example: 2.1744
        placeId:
          type: string
          description: Google Places ID
          example: "ChIJ5TCOcRaYpBIRCmZHTz37sEQ"
    PriceDto:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: double
          minimum: 0
          description: Price amount
          example: 25.50
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: Currency code (ISO 4217)
          example: "EUR"
        per:
          type: string
          description: Price unit (per person, per night, etc.)
          example: "person"

    # Agent DTOs
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "INTERNAL_ERROR"
        message:
          type: string
          description: Error message
          example: "Internal server error"
        hint:
          type: string
          description: Error hint
          example: "An unexpected error occurred"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path
          example: "/api/v1/itineraries"
  securitySchemes:
    # No authentication required in development mode
    # This can be extended for production use
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (not currently required)
security:
  # No security requirements in development mode
  - {}